As mentioned, using OcelotLogger instead of AspDotNeLogger as default logger name
