+++ src/Ocelot/Logging/OcelotLoggerFactory.cs
@@ -3,12 +3,12 @@ using Ocelot.Infrastructure.RequestData;
 
 namespace Ocelot.Logging
 {
-    public class AspDotNetLoggerFactory : IOcelotLoggerFactory
+    public class OcelotLoggerFactory : IOcelotLoggerFactory
     {
         private readonly ILoggerFactory _loggerFactory;
         private readonly IRequestScopedDataRepository _scopedDataRepository;
 
-        public AspDotNetLoggerFactory(ILoggerFactory loggerFactory, IRequestScopedDataRepository scopedDataRepository)
+        public OcelotLoggerFactory(ILoggerFactory loggerFactory, IRequestScopedDataRepository scopedDataRepository)
         {
             _loggerFactory = loggerFactory;
             _scopedDataRepository = scopedDataRepository;
@@ -17,7 +17,7 @@ namespace Ocelot.Logging
         public IOcelotLogger CreateLogger<T>()
         {
             var logger = _loggerFactory.CreateLogger<T>();
-            return new AspDotNetLogger(logger, _scopedDataRepository);
+            return new OcelotLogger(logger, _scopedDataRepository);
         }
     }
 }
